add parlay files

+++ tree.cpp
@@ -112,7 +112,7 @@ void print_tree(vector<vector<int>> &adj, int n)
     for (int i = 0; i < n; i++)
     {
         cout << i << ": ";
-        for (int j = 0; j < adj[i].size(); j++)
+        for (int j = 0; j < (int) adj[i].size(); j++)
         {
             cout << adj[i][j] << ",";
         }
